refactor: decompose command_handlers.rs into directory module (#435)
Split 716 LOC command_handlers.rs into 4 focused sub-modules:
- mod.rs (~280 LOC): dispatch, exit, compact, help, custom commands
- stats_handlers.rs (~160 LOC): /stats, /cost, /context, /usage
- session_handlers.rs (~140 LOC): /save, /load, /sessions
- utility_handlers.rs (~60 LOC): /bashes

Re-exports from mod.rs preserve all existing call sites unchanged.
